dag: refactoring, take PipelineTask instead of the whole Pipeline ⛏
… and rename the package from resources to dag
